AmpliVision/requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
matplotlib==3.9.2
pillow-heif==0.18.0
pillow==10.4.0
networkx==3.3
pandas==2.2.2
scipy==1.14.1
pyod==2.0.2
tensorflow==2.15
keras==2.15.0
tensorflow-addons==0.23
seaborn==0.13.2
mahotas==1.4.18
pydicom